Refactor tests for HttpSamplingManager

+++ jaeger-core/src/main/java/com/uber/jaeger/samplers/http/RateLimitingSamplingStrategy.java
@@ -25,5 +25,5 @@ import lombok.Value;
 
 @Value
 public class RateLimitingSamplingStrategy {
-  int maxTracesPerSecond;
+  double maxTracesPerSecond;
 }
